LayoutTests/accessibility/heading-level-expected.txt

This tests that headings have a level matching the implicit default value or explicitly defined aria-level value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS axElement.intValue is parseInt(el.getAttribute('data-expected'))
PASS successfullyParsed is true

TEST COMPLETE
